Crate cargo_ghp_upload

Reexports

extern crate std as std;

extern crate quicli;

extern crate fs_extra;

use std::prelude::v1::*;

use quicli::prelude::*;

use std::env;

use std::fs;

use std::str;

use std::ffi::OsString;

use std::path::Path;

use std::path::PathBuf;

use std::process::Command;

use std::process::ExitStatus;

Macros

bail
Exits a function early with an Error.

debug
Logs a message at the debug level.

ensure
Exits a function early with an Error if the condition is not satisfied.

error
Logs a message at the error level.

format_err
Constructs an Error using the standard string interpolation syntax.

info
Logs a message at the info level.

log
The standard logging macro.

log_enabled
Determines if a message logged at the specified level in that module will be logged.

trace
Logs a message at the trace level.

warn
Logs a message at the warn level.

Structs

Args
Upload documentation straight to GitHub Pages while maintaining branch separation and history

Context

Enums

CargoArgs

Functions

get_context

ghp_upload

main

require_success

run